mcp-servers/slack-mcp/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
slack-sdk==3.33.4
google-cloud-secret-manager==2.20.0
